services/filestorage/filestorage-api/requirements.txt
fastapi
uvicorn
itsdangerous
aiofiles
python-multipart
pyjwt[crypto]
httpx
cryptography